requirements.txt
openai~=1.3.6
flask~=2.2.2
python-dotenv~=0.21.0